# openapi-resolved.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Location API",
    "description": "Location service endpoints for university project",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://location-app:8080/api",
      "description": "Location service (use service name, not pod name)"
    }
  ],
  "tags": [
    {
      "name": "location",
      "description": "APIs for location services"
    }
  ],
  "paths": {
    "/location": {
      "post": {
        "tags": [
          "location"
        ],
        "summary": "Create a new location",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/location_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Location created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/location/{id}": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Get location by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the location",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "404": {
            "description": "Location not found"
          }
        }
      },
      "put": {
        "tags": [
          "location"
        ],
        "summary": "Update location fully",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the location",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/location_id_body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Location updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Location not found"
          }
        }
      },
      "delete": {
        "tags": [
          "location"
        ],
        "summary": "Delete location by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the location",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Location deleted"
          },
          "404": {
            "description": "Location not found"
          }
        }
      },
      "patch": {
        "tags": [
          "location"
        ],
        "summary": "Update location partially",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the location",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/location_id_body_1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Location updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "404": {
            "description": "Location not found"
          }
        }
      }
    },
    "/location/distance/{id}": {
      "get": {
        "tags": [
          "location"
        ],
        "summary": "Get distance from a reference point to a specific location",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the location",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "refLat",
            "in": "query",
            "description": "Reference point latitude",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "refLng",
            "in": "query",
            "description": "Reference point longitude",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200"
                }
              }
            }
          },
          "404": {
            "description": "Location not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Location": {
        "required": [
          "id",
          "lat",
          "lng"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 123
          },
          "lat": {
            "type": "number",
            "format": "double",
            "example": 46.0569
          },
          "lng": {
            "type": "number",
            "format": "double",
            "example": 14.5058
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "location_body": {
        "required": [
          "lat",
          "lng"
        ],
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "format": "double",
            "example": 46.0569
          },
          "lng": {
            "type": "number",
            "format": "double",
            "example": 14.5058
          }
        }
      },
      "location_id_body": {
        "required": [
          "lat",
          "lng"
        ],
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lng": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "location_id_body_1": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lng": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "inline_response_200": {
        "type": "object",
        "properties": {
          "distance": {
            "type": "number",
            "description": "Distance in kilometers",
            "format": "double"
          },
          "unit": {
            "type": "string",
            "example": "km"
          }
        }
      }
    }
  }
}